Fix UI slips. Fix syncing issue. V 1.0.1
The syncing issue was due to the playback time not being set on the
host device in the -startplaybackAtTime method. We were relying on it
already being taken from the device. However, when a new song started,
the argument was manually passed at 0, and due to another minor bug
that caused the next song to play prematurely, the host was always a
second or so off.